Time: 11 ms (39.53%) | Memory: 23 MB (42.16%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -5,10 +5,10 @@ public:
         int end = *max_element(piles.begin(), piles.end());
         int res = end;
         while (start <= end) {
-            long long mid = start + (end - start) / 2;
+            int mid = start + (end - start) / 2;
             long long total = 0;
             for (size_t i = 0; i < piles.size(); ++i) {
-                total += ((long long)piles[i] + mid - 1) / mid;
+                total += (piles[i] + mid - 1) / mid;
                 if (total > h) {
                     break;
                 }
